refactor(opengl): reorganize and relocate OpenGL utilities
