+++ app/src/main/java/com/github/lookupgroup27/lookup/util/opengl/BufferUtils.kt
@@ -1,4 +1,4 @@
-package com.github.lookupgroup27.lookup.util
+package com.github.lookupgroup27.lookup.util.opengl
 
 import java.nio.ByteBuffer
 import java.nio.ByteOrder
