+++ app/src/main/java/com/github/lookupgroup27/lookup/util/opengl/ShaderUtils.kt
@@ -1,4 +1,4 @@
-package com.github.lookupgroup27.lookup.util
+package com.github.lookupgroup27.lookup.util.opengl
 
 import android.content.Context
 import android.util.Log
